+++ admin/src/main/java/com/wuyan/masteryi/admin/service/GoodsService.java
@@ -25,7 +25,7 @@ public interface GoodsService {
                 String goodsCoverUrl, Integer collectNum, Integer sellNum);
     Map<String, Object> addSpecs(Integer goodsId, String specs, Integer stock, float price);
     Map<String, Object> changeStock(Integer newStock, Integer goodSpecsId);
-    Map<String, Object> changePrice(Integer newPrice, Integer goodSpecsId);
+    Map<String, Object> changePrice(float newPrice, Integer goodSpecsId);
     Map<String, Object> deleteSpecs(Integer goodSpecsId);
     Map<String, Object> deleteGoods(Integer goodsId);
     Map<String,Object> getSpecsDesc(int id);
